# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bridge-Watch API",
    "description": "Real-time Stellar bridge monitoring API. Provides bridge statuses, asset health, liquidity, price feeds, alert management, analytics, and administrative controls.\n\n## Authentication\nProtected endpoints require an API key in the `x-api-key` header.\n\n## Rate Limiting\nAll endpoints are rate-limited per IP using a sliding-window algorithm backed by Redis. Exceeding the limit returns **429 Too Many Requests** with a `Retry-After` header.\n\n## Versioning\nAll REST endpoints are prefixed with `/api/v1/`. Breaking changes increment the version segment.\n\n## Error Format\n```json\n{ \"error\": \"Short label\", \"message\": \"Human-readable description\" }\n```",
    "version": "1.0.0",
    "contact": {
      "name": "Bridge-Watch Team",
      "url": "https://github.com/StellaBridge/Bridge-Watch"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" },
    { "url": "https://api.bridge-watch.io", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for protected endpoints"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/alerts/rules": {
      "get": {
        "tags": ["Alerts"],
        "summary": "List alert rules for an owner",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "name": "owner", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Alert rules" },
          "400": { "description": "Missing owner param" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Alerts"],
        "summary": "Create an alert rule",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ownerAddress", "name", "assetCode", "conditions"],
                "properties": {
                  "ownerAddress": { "type": "string" },
                  "name": { "type": "string", "example": "USDC health drop" },
                  "assetCode": { "type": "string", "example": "USDC" },
                  "conditions": { "type": "array", "items": { "type": "object" } },
                  "conditionOp": { "type": "string", "enum": ["AND", "OR"], "default": "AND" },
                  "priority": { "type": "string", "enum": ["low", "medium", "high", "critical"], "default": "medium" },
                  "cooldownSeconds": { "type": "integer", "default": 300 },
                  "webhookUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Rule created" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/alerts/rules/{ruleId}": {
      "get": {
        "tags": ["Alerts"],
        "summary": "Get a single alert rule",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "name": "ruleId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Alert rule" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Rule not found" }
        }
      },
      "patch": {
        "tags": ["Alerts"],
        "summary": "Update an alert rule",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "name": "ruleId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": { "description": "Updated rule" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Rule not found" }
        }
      },
      "delete": {
        "tags": ["Alerts"],
        "summary": "Delete an alert rule",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "name": "ruleId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["ownerAddress"], "properties": { "ownerAddress": { "type": "string" } } } } }
        },
        "responses": {
          "204": { "description": "Rule deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Rule not found" }
        }
      }
    },
    "/api/v1/alerts/rules/{ruleId}/active": {
      "patch": {
        "tags": ["Alerts"],
        "summary": "Pause or resume an alert rule",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "name": "ruleId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["ownerAddress", "isActive"], "properties": { "ownerAddress": { "type": "string" }, "isActive": { "type": "boolean" } } } } }
        },
        "responses": { "200": { "description": "Success" }, "401": { "description": "Unauthorized" }, "404": { "description": "Rule not found" } }
      }
    },
    "/api/v1/alerts/rules/{ruleId}/events": {
      "get": {
        "tags": ["Alerts"],
        "summary": "Get events fired by a specific rule",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "ruleId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": { "200": { "description": "Rule events" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/api/v1/alerts/rules/bulk": {
      "post": {
        "tags": ["Alerts"],
        "summary": "Bulk create alert rules",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "201": { "description": "Rules created" }, "401": { "description": "Unauthorized" } }
      },
      "patch": {
        "tags": ["Alerts"],
        "summary": "Bulk update alert rules",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Rules updated" }, "401": { "description": "Unauthorized" } }
      },
      "delete": {
        "tags": ["Alerts"],
        "summary": "Bulk delete alert rules",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Rules deleted" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/api/v1/alerts/history": {
      "get": {
        "tags": ["Alerts"],
        "summary": "Get paginated alert history",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "Paginated alert history" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/api/v1/alerts/history/{assetCode}": {
      "get": {
        "tags": ["Alerts"],
        "summary": "Get alert history for an asset",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "assetCode", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": { "200": { "description": "Asset alert history" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/api/v1/alerts/stats": {
      "get": {
        "tags": ["Alerts"],
        "summary": "Get alert statistics for an owner",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "name": "owner", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Alert stats" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/api/v1/alerts/recent": {
      "get": {
        "tags": ["Alerts"],
        "summary": "Get most recent alert events",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } }],
        "responses": { "200": { "description": "Recent events" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/api/v1/alerts/test": {
      "post": {
        "tags": ["Alerts"],
        "summary": "Dry-run an alert rule against current metrics",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["rule", "metrics"], "properties": { "rule": { "type": "object" }, "metrics": { "type": "object" } } } } }
        },
        "responses": { "200": { "description": "Dry-run result" }, "401": { "description": "Unauthorized" } }
      }
    },
...
  }
}